refactor: centralize routing and dependency injection
