+++ src/interfaces/IStorageRepository.ts
@@ -3,5 +3,5 @@ import { ProjectSettings } from '../models/ProjectSettings';
 export interface IStorageRepository {
   createNewProjectFromTemplate(
     projectSettings: ProjectSettings
-  ): Promise<{ projectId: string }>;
+  ): Promise<{ projectId: string; downloadUrl: string }>;
 }
