+++ src/repositories/index.ts
@@ -1,0 +1,2 @@
+export * from './StorageRepository';
+export * from './projectRepository';
